Move ITA ops to a package
