+++ Modern/ops/src/main/java/org/bonej/ops/ita/NPoint.java
@@ -1,4 +1,4 @@
-package org.bonej.ops;
+package org.bonej.ops.ita;
 
 import java.util.List;
