+++ Modern/ops/src/test/java/org/bonej/ops/ita/CleanShortEdgesTest.java
@@ -1,8 +1,8 @@
 
-package org.bonej.ops;
+package org.bonej.ops.ita;
 
-import static org.bonej.ops.CleanShortEdges.findClusters;
-import static org.bonej.ops.CleanShortEdges.isShortEdge;
+import static org.bonej.ops.ita.CleanShortEdges.findClusters;
+import static org.bonej.ops.ita.CleanShortEdges.isShortEdge;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
@@ -22,6 +22,8 @@ import net.imagej.ImageJ;
 import net.imagej.ops.special.function.BinaryFunctionOp;
 import net.imagej.ops.special.function.Functions;
 
+import org.bonej.ops.TestGraphs;
+import org.bonej.ops.ita.CleanShortEdges;
 import org.bonej.utilities.GraphUtil;
 import org.junit.After;
 import org.junit.Before;
